# book-stork-client/doc/openapi.yaml
openapi: 3.0.1
info:
  title: BookStork API
  description: API con Arquitectura Hexagonal + DDD
  version: v1
paths:
...
  /api/v1/loans:
    get:
      tags:
        - Loans
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanDtoPagedResult'
    post:
      tags:
        - Loans
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLoanRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateLoanRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateLoanRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanDto'
  /api/v1/loans/me:
    get:
      tags:
        - Loans
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanDtoPagedResult'
  '/api/v1/loans/{id}':
    get:
      tags:
        - Loans
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanDto'
  '/api/v1/loans/{id}/return':
    post:
      tags:
        - Loans
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoanDto'
  /api/v1/reservations:
    get:
      tags:
        - Reservations
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDtoPagedResult'
    post:
      tags:
        - Reservations
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateReservationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDto'
  /api/v1/reservations/me:
    get:
      tags:
        - Reservations
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDtoPagedResult'
  '/api/v1/reservations/{id}/cancel':
    post:
      tags:
        - Reservations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDto'
  /api/v1/users:
    get:
      tags:
        - Users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDtoPagedResult'
  '/api/v1/users/{id}':
    get:
      tags:
        - Users
      operationId: GetUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
    delete:
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
...
components:
  schemas:
...
    BookSummaryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        isbn:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        authorNames:
          type: string
          nullable: true
        categoryName:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        averageRating:
          type: number
          format: double
        language:
          type: string
          nullable: true
      additionalProperties: false
...
    CreateLoanRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        bookId:
          type: string
          format: uuid
        days:
          type: integer
          format: int32
      additionalProperties: false
    CreateReservationRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        bookId:
          type: string
          format: uuid
      additionalProperties: false
...
    LoanDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        userFullName:
          type: string
          nullable: true
        book:
          $ref: '#/components/schemas/BookSummaryDto'
        status:
          type: string
          nullable: true
        loanedAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        returnedAt:
          type: string
          format: date-time
          nullable: true
        isOverdue:
          type: boolean
        daysRemaining:
          type: integer
          format: int32
      additionalProperties: false
    LoanDtoPagedResult:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LoanDto'
          nullable: true
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalCount:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
          readOnly: true
        hasNextPage:
          type: boolean
          readOnly: true
        hasPreviousPage:
          type: boolean
          readOnly: true
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    ReservationDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        userFullName:
          type: string
          nullable: true
        book:
          $ref: '#/components/schemas/BookSummaryDto'
        status:
          type: string
          nullable: true
        reservedAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        fulfilledAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
    ReservationDtoPagedResult:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReservationDto'
          nullable: true
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalCount:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
          readOnly: true
        hasNextPage:
          type: boolean
          readOnly: true
        hasPreviousPage:
          type: boolean
          readOnly: true
      additionalProperties: false
...
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        notificationPreference:
          type: string
          nullable: true
      additionalProperties: false
...
    UserDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          nullable: true
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        fullName:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        loanLimit:
          type: integer
          format: int32
        activeLoansCount:
          type: integer
          format: int32
        notificationPreference:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
    UserDtoPagedResult:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserDto'
          nullable: true
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalCount:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
          readOnly: true
        hasNextPage:
          type: boolean
          readOnly: true
        hasPreviousPage:
          type: boolean
          readOnly: true
      additionalProperties: false
...